--- query.txt
@@ -1,4 +1,0 @@
-tshark -i wlp0s20f3 -f "(udp src port 53) or (udp src port 5353)" -n -T fields -e frame.interface_name -e _ws.col.Time -e ip.dst -e udp.dstport -e ip.src -e udp.srcport -e dns.flags.rcode -e dns.qry.name -e dns.qry.name.len -e dns.count.labels -e dns.qry.type -e dns.resp.name -e dns.resp.type -e dns.resp.ttl -e dns.resp.len -e dns.rr.udp_payload_size -e dns.a -e dns.aaaa -e dns.soa.rname -e dns.soa.mname -e dns.txt -E header=y -E separator="," -E occurrence=f -E aggregator="|" > output.csv
-
-tshark -i [interface] -Y "dns.qry.type == 1 or dns.qry.type == 28 or dns.qry.type == 2 or dns.qry.type == 12 or dns.qry.type == 6 or dns.qry.type == 16 or dns.qry.type == 255" -T fields -e frame.number -e ip.src -e dns.qry.name -e dns.qry.type -e dns.qry.class -e dns.flags.response -e dns.resp.type -e dns.resp.class -E header=y -E separator=, -E quote=d -E occurrence=f > dns_queries.csv
-
