+++ tsharkPackages/tsharkServices/atf_dns_data_insertion.service
@@ -1,0 +1,15 @@
+[Unit]
+Description=ATF DNS Data Insertion Service
+After=network.target
+StartLimitIntervalSec=0
+
+[Service]
+Type=simple
+Restart=always
+RestartSec=300
+User=attackfence
+WorkingDirectory=/opt/attackfence/NDR/tsharkQueryData/
+ExecStart=/usr/bin/python3 /opt/attackfence/NDR/tsharkQueryData/analyseDNSData.py
+
+[Install]
+WantedBy=multi-user.target
